loader: Make logging and getenv their own headers
This commit pulls the logging capabilities into a single log.h/.c file and does
the same for the getenv functionality.
