+++ loader/asm_offset.c
@@ -23,6 +23,7 @@
 
 #include <stdio.h>
 #include "loader.h"
+#include "log.h"
 
 #if !defined(_MSC_VER) || (_MSC_VER >= 1900)
 #define SIZE_T_FMT "%-8zu"
