+++ loader/loader_common.h
@@ -31,16 +31,6 @@
 
 #include "vk_loader_platform.h"
 
-enum vulkan_loader_debug_flags {
-    VULKAN_LOADER_INFO_BIT = 0x01,
-    VULKAN_LOADER_WARN_BIT = 0x02,
-    VULKAN_LOADER_PERF_BIT = 0x04,
-    VULKAN_LOADER_ERROR_BIT = 0x08,
-    VULKAN_LOADER_DEBUG_BIT = 0x10,
-    VULKAN_LOADER_LAYER_BIT = 0x20,
-    VULKAN_LOADER_IMPLEMENTATION_BIT = 0x40,
-};
-
 enum layer_type_flags {
     VK_LAYER_TYPE_FLAG_INSTANCE_LAYER = 0x1,  // If not set, indicates Device layer
     VK_LAYER_TYPE_FLAG_EXPLICIT_LAYER = 0x2,  // If not set, indicates Implicit layer
